+++ docs/changelog.rst
@@ -2,6 +2,7 @@
 Changelog
 =========
 
+- :feature:`-` Add Codecov support to ``pytest.coverage``.
 - :support:`-` Rely on Invoke 1.6+ for some of its new features.
 - :feature:`-` ``packaging.release.prepare`` grew a ``dry_run`` flag to match
   the rest of its friends.